# openapi/manager/openapi.yaml
openapi: 3.0.3
info:
  title: manager
  description: ''
  license:
    name: ''
  version: 0.1.0
paths:
...
  /v1/functions:
    get:
      tags:
      - Functions
      operationId: list
      responses:
        '200':
          description: Functions listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListFunctionsResp'
        '500':
          description: Failed to list functions
    post:
      tags:
      - Functions
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFunctionReq'
        required: true
      responses:
        '200':
          description: Function created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateFunctionResp'
        '400':
          description: Invalid request
        '500':
          description: Failed to create function
  /v1/functions/{id}:
    get:
      tags:
      - Functions
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Function fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFunctionResp'
        '404':
          description: Function not found
        '500':
          description: Failed to fetch function
    put:
      tags:
      - Functions
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFunctionReq'
        required: true
      responses:
        '200':
          description: Function updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFunctionResp'
        '404':
          description: Function not found
        '500':
          description: Failed to update function
    delete:
      tags:
      - Functions
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Function deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '404':
          description: Function not found
        '500':
          description: Failed to delete function
  /v1/functions/{id}/invoke:
    post:
      tags:
      - Functions
      operationId: invoke
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvokeFunctionReq'
        required: true
      responses:
        '200':
          description: Function invoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvokeFunctionResp'
        '404':
          description: Function not found
        '500':
          description: Failed to invoke function
  /v1/functions/{id}/logs:
    get:
      tags:
      - Functions
      operationId: logs
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: status
        in: path
        required: true
        schema:
          type: string
          nullable: true
      - name: limit
        in: path
        required: true
        schema:
          type: integer
          format: int64
          nullable: true
      responses:
        '200':
          description: Invocation logs fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListInvocationsResp'
        '404':
          description: Function not found
        '500':
          description: Failed to fetch logs
...
  /v1/snapshots/{id}/instantiate:
    post:
      tags:
      - Snapshots
      operationId: instantiate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        description: Optional overrides when instantiating a snapshot
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstantiateSnapshotReq'
        required: true
      responses:
        '200':
          description: Snapshot instantiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstantiateSnapshotResp'
        '404':
          description: Snapshot not found
        '502':
          description: Failed to instantiate snapshot
  /v1/templates:
    get:
      tags:
      - Templates
      operationId: list
      responses:
        '200':
          description: Templates listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTemplatesResp'
        '500':
          description: Failed to list templates
    post:
      tags:
      - Templates
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateReq'
        required: true
      responses:
        '200':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTemplateResp'
        '500':
          description: Failed to create template
  /v1/templates/{id}:
    get:
      tags:
      - Templates
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Template fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTemplateResp'
        '404':
          description: Template not found
        '500':
          description: Failed to fetch template
    put:
      tags:
      - Templates
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTemplateReq'
        required: true
      responses:
        '200':
          description: Template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateTemplateResp'
        '404':
          description: Template not found
        '500':
          description: Failed to update template
    delete:
      tags:
      - Templates
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Template deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '404':
          description: Template not found
        '500':
          description: Failed to delete template
  /v1/templates/{id}/instantiate:
    post:
      tags:
      - Templates
      operationId: instantiate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstantiateTemplateReq'
        required: true
      responses:
        '200':
          description: Template instantiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstantiateTemplateResp'
        '404':
          description: Template not found
        '500':
          description: Failed to instantiate template
...
components:
  schemas:
...
    CreateFunctionReq:
      type: object
      required:
      - name
      - runtime
      - code
      - handler
      properties:
        code:
          type: string
        env_vars:
          nullable: true
        handler:
          type: string
        memory_mb:
          type: integer
          format: int32
        name:
          type: string
        runtime:
          type: string
        timeout_seconds:
          type: integer
          format: int32
        vcpu:
          type: integer
          format: int32
    CreateFunctionResp:
      type: object
      required:
      - id
      properties:
        id:
          type: string
          format: uuid
...
    CreateTemplateReq:
      type: object
      required:
      - name
      - spec
      properties:
        name:
          type: string
        spec:
          $ref: '#/components/schemas/TemplateSpec'
    CreateTemplateResp:
      type: object
      required:
      - id
      properties:
        id:
          type: string
          format: uuid
...
    Function:
      type: object
      required:
      - id
      - name
      - runtime
      - code
      - handler
      - timeout_seconds
      - memory_mb
      - vcpu
      - port
      - state
      - created_at
      - updated_at
      properties:
        code:
          type: string
        created_at:
          type: string
          format: date-time
        created_by_user_id:
          type: string
          format: uuid
          nullable: true
        env_vars:
          nullable: true
        guest_ip:
          type: string
          nullable: true
        handler:
          type: string
        id:
          type: string
          format: uuid
        last_invoked_at:
          type: string
          format: date-time
          nullable: true
        memory_mb:
          type: integer
          format: int32
        name:
          type: string
        port:
          type: integer
          format: int32
        runtime:
          type: string
        state:
          type: string
        timeout_seconds:
          type: integer
          format: int32
        updated_at:
          type: string
          format: date-time
        vcpu:
          type: integer
          format: int32
        vm_id:
          type: string
          format: uuid
          nullable: true
    FunctionInvocation:
      type: object
      required:
      - id
      - function_id
      - status
      - duration_ms
      - request_id
      - event
      - invoked_at
      properties:
        duration_ms:
          type: integer
          format: int64
        error:
          type: string
          nullable: true
        event: {}
        function_id:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        invoked_at:
          type: string
          format: date-time
        logs:
          type: array
          items:
            type: string
        memory_used_mb:
          type: integer
          format: int32
          nullable: true
        request_id:
          type: string
        response:
          nullable: true
        status:
          type: string
...
    GetFunctionResp:
      type: object
      required:
      - item
      properties:
        item:
          $ref: '#/components/schemas/Function'
...
    GetTemplateResp:
      type: object
      required:
      - item
      properties:
        item:
          $ref: '#/components/schemas/Template'
...
    InstantiateSnapshotReq:
      type: object
      properties:
        name:
          type: string
          nullable: true
    InstantiateSnapshotResp:
      type: object
      required:
      - id
      - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    InstantiateTemplateReq:
      type: object
      required:
      - name
      properties:
        name:
          type: string
    InstantiateTemplateResp:
      type: object
      required:
      - id
      properties:
        id:
          type: string
          format: uuid
    InvokeFunctionReq:
      type: object
      required:
      - event
      properties:
        event: {}
    InvokeFunctionResp:
      type: object
      required:
      - request_id
      - status
      - duration_ms
      properties:
        duration_ms:
          type: integer
          format: int64
        error:
          type: string
          nullable: true
        logs:
          type: array
          items:
            type: string
        request_id:
          type: string
        response:
          nullable: true
        status:
          type: string
...
    ListFunctionsResp:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Function'
...
    ListInvocationsResp:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FunctionInvocation'
...
    ListTemplatesResp:
      type: object
      required:
      - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Template'
...
    OkResponse:
      type: object
      required:
      - ok
      properties:
        ok:
          type: boolean
...
    TemplateSpec:
      type: object
      required:
      - vcpu
      - mem_mib
      properties:
        kernel_image_id:
          type: string
          format: uuid
          nullable: true
        kernel_path:
          type: string
          nullable: true
        mem_mib:
          type: integer
          format: int32
          minimum: 0
        rootfs_image_id:
          type: string
          format: uuid
          nullable: true
        rootfs_path:
          type: string
          nullable: true
        rootfs_size_mb:
          type: integer
          format: int32
          nullable: true
          minimum: 0
        vcpu:
          type: integer
          format: int32
          minimum: 0
...
    UpdateFunctionReq:
      type: object
      properties:
        code:
          type: string
          nullable: true
        env_vars:
          nullable: true
        handler:
          type: string
          nullable: true
        memory_mb:
          type: integer
          format: int32
          nullable: true
        name:
          type: string
          nullable: true
        runtime:
          type: string
          nullable: true
        timeout_seconds:
          type: integer
          format: int32
          nullable: true
...
    UpdateTemplateReq:
      type: object
      required:
      - name
      - spec
      properties:
        name:
          type: string
        spec:
          $ref: '#/components/schemas/TemplateSpec'
    UpdateTemplateResp:
      type: object
      required:
      - item
      properties:
        item:
          $ref: '#/components/schemas/Template'
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your Bearer token from /v1/auth/login
security:
- bearer_auth: []
...